silence warnings

+++ src/GearsystemCore.cpp
@@ -465,11 +465,11 @@ bool GearsystemCore::SaveState(std::ostream& stream, size_t& size)
         //m_pAudio->SaveState(stream);
         //m_pMemory->GetCurrentRule()->SaveState(stream);
 
-        size = stream.tellp();
+        size = static_cast<size_t>(stream.tellp());
         size += (sizeof(u32) * 2);
 
         u32 header_magic = GS_SAVESTATE_MAGIC;
-        u32 header_size = size;
+        size_t header_size = size;
 
         stream.write(reinterpret_cast<const char*> (&header_magic), sizeof(header_magic));
         stream.write(reinterpret_cast<const char*> (&header_size), sizeof(header_size));
@@ -562,7 +562,7 @@ bool GearsystemCore::LoadState(std::istream& stream)
         u32 header_size = 0;
 
         stream.seekg(0, ios::end);
-        size_t size = stream.tellg();
+        size_t size = static_cast<size_t>(stream.tellg());
         stream.seekg(0, ios::beg);
 
         stream.seekg(-2 * (sizeof(u32)), ios::end);
